refactor: move round df fn to core utils
